Fixed postAuthentication() bug
postAuthentication() method is run after login for every registered provider which meant that it also ran if the LocalPasswordPrimaryAuthenticationProvider was that one who authenticated a user. postAuthentication() then tried to access something that was not there. This is now fixed. Post auth tasks will only be performed if the ExternalDatabaseAuth provider was the one that successfully authenticated.